Move GormRepository to repository.go
Previous file division added an additional step when trying to look up
an implementation of a method. You would first be redirected to
repository.go and then it would require a manual switch to the
gorm_repository.go file. Moving GormRepository to the same file allows
for an easier transition.